Create CLAHE Technique
Contrast limited adaptive histogram equalization (CLAHE) is used to improve the visibility level of foggy image or video. In this code, I used CLAHE enhancement method for improving the Image quality in the real-time system. Adaptive histogram equalization (AHE) is different from normal histogram equalization because AHE use several methods each corresponding to different parts of image and used them to redistribute the lightness value of the image and in case of CLAHE `Distribution' parameter are used to define the shape of histogram which produces the better quality result compare then adaptive histogram equalization (AHE). The drawback of AHE is work over homogeneous fog but CLAHE applied over both homogeneous and heterogeneous fog and single image and video system. And the second drawback of AHE is used `cumulation function' which applied to the only grey level image but CLAHE used both images coloured and grey levels.